refactor: extract release package preflight
